requirements/base.txt
pytz==2018.7  # https://github.com/stub42/pytz
python-slugify==1.2.6  # https://github.com/un33k/python-slugify
Pillow==5.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.2  # https://github.com/evansd/whitenoise
redis>=2.10.6, < 3  # pyup: < 3 # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==2.0.9  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.38.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Third party packages
# ------------------------------------------------------------------------------
django-braces==1.13.0
